pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aiolymp-hometask"
version = "1.0.0"
description = "Kazakhstan nationals olympiad on AI Hometask: Structured Output with LoRA"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.1.0",
    "transformers>=4.46.0",
    "peft>=0.13.0",
    "bitsandbytes>=0.44.0",
    "datasets>=3.0.0",
    "accelerate>=1.0.0",
    "pyyaml>=6.0",
    "tomli>=2.0.0;python_version<'3.11'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.setuptools.packages.find]
include = ["*"]
